# apps/backend/openapi/v1.yaml
...
openapi: 3.1.1
info:
  title: Argentum API
  # A date rather than a semver, matching what `GET /v1/me` reports as
  # `api_version` — "you are on 2026-07-30" locates a deploy, where "1.0.0"
  # locates nothing. A test pins this equal to `apiv1.Version`.
  version: '2026-07-30'
  summary: Ask for a report or an answer over HTTP.
  description: |
    Argentum turns a question about a company's own data into an answer or a
    branded document. This is the machine-facing half: the same agent the
    dashboard drives, reachable with an API key.

    ## The two products

    **Reports.** `POST /v1/reports/render` takes a document spec and returns a
    file — deterministic, sub-second, no LLM. `POST /v1/reports` takes a prompt
    and runs a real agent turn that ends by generating one.

    **Chat.** `POST /v1/chat` takes a question and returns an answer, either
    streamed over server-sent events or waited for on the connection.

    ## The contract every route inherits

    * **Authentication** is `Authorization: Bearer arg_…`. A dashboard session
      is refused here, and a key is refused on the dashboard's own `/api`
      routes: they are different authorities.
    * **Scopes** are fixed when a key is minted and cannot be changed. Each
      operation below names the one it requires under `x-argentum-scope`.
    * **Every response carries `X-Request-Id`**, including failures. Quote it
      in a support conversation.
...
      `error.message` is for a human.
...

    ## Streams

...

...

...
servers:
...
security:
  - apiKey: []
...
paths:
...
  /v1/reports:
    post:
      operationId: createReport
      tags: [Reports]
      summary: Ask an agent for a report
      description: |
        A prompt in, a real agent turn behind it, a document at the end. Seconds
        to minutes, priced in tokens.

        Three ways to collect the result, and a caller picks one: poll
        `GET /v1/reports/{id}`, stream `GET /v1/reports/{id}/events`, or supply
        a `callback_url` and verify the signature described by
        `GET /v1/me`'s `webhooks` block.
      x-argentum-scope: write:reports
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateReportRequest' }
            example:
              prompt: Monthly revenue for 2024, with a bar chart by month and a short commentary.
              format: pdf
              user_ref: u_42
      responses:
        '202':
          description: The job was accepted. It is `queued` until a worker picks it up.
          headers:
            X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
            Idempotent-Replay: { $ref: '#/components/headers/Idempotent-Replay' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Report' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '402': { $ref: '#/components/responses/BudgetExhausted' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/ServerError' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /v1/reports/{id}:
    get:
      operationId: getReport
      tags: [Reports]
      summary: Poll a report
      description: |
        Once `status` is `completed` the `document` block carries a freshly
        presigned `download_url`. It is re-issued on every read rather than
        stored, so a caller who saved one and came back an hour later gets a
        working link without paying to regenerate the document.
      x-argentum-scope: read:documents
      parameters:
        - $ref: '#/components/parameters/ReportID'
      responses:
        '200':
          description: The report, and its document once there is one.
          headers:
            X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Report' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
  /v1/threads:
    get:
      operationId: listThreads
      tags: [Chat]
      summary: List the conversations this integration started
      description: |
        `api` threads only. A machine credential that could list the dashboard's
        threads would let a leaked key read the staff's own chat history.
      x-argentum-scope: read:threads
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: user_ref
          in: query
          description: Only threads belonging to this end user.
          required: false
          schema: { type: string }
      responses:
        '200':
          description: One page of threads, newest first.
          headers:
            X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ThreadPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/ServerError' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      description: |
        `Authorization: Bearer arg_<prefix>_<secret>`, minted in the dashboard
        under Settings → API Keys. The secret is shown once.

        **Header only.** The dashboard's own token may travel in a cookie or a
        query parameter because a browser cannot set a header on a WebSocket
        upgrade; neither applies to a machine caller, and both are how a
        credential ends up in an access log or a proxy trace.
  parameters:
    Limit:
      name: limit
      in: query
      description: |
        Page size. A value over the repository's cap is clamped rather than
        refused — a caller asking for 10 000 rows wants as many as they can get,
        and the cursor is what makes the rest reachable.
      required: false
      schema: { type: integer, minimum: 1, examples: [50] }
    Cursor:
      name: cursor
      in: query
      description: |
        The `next_cursor` from the previous page. Opaque: it encodes the
        `(created_at, id)` of the last row you saw, and hand-building one is a
        `400`, not a silent restart at page one.
      required: false
      schema: { type: string }
    ReportID:
      name: id
      in: path
      required: true
      description: A report id.
      schema: { type: string, format: uuid }
...
    IdempotencyKeyRequired:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        A unique string per logical request; a UUID is the usual choice. Both
        SDKs generate one automatically when you do not supply it.

        A second request under the same key ends one of three ways:

        * the original finished — you get its result again, with
          `Idempotent-Replay: true`, and nothing is billed twice;
        * the original is still running — `409 request_in_flight`, carrying the
          ids of the work you are already waiting on;
        * the body differs from the original's — `409 idempotency_key_reuse`.
          Use a new key for a new request.

        A request that fails forgets its key, so a retry after a 500 is a fresh
        attempt. The one exception is `POST /v1/chat`'s 504, where the request
        failed and the work did not.
      schema: { type: string, maxLength: 255 }
...
  headers:
    X-Request-Id:
      description: |
        This request's id, on every response including failures. Send your own
        and it is echoed back, provided it is at most 128 characters of
        `[A-Za-z0-9_.:-]`.
      schema: { type: string, examples: ['req_1945841125adc2c7bb618224f5fa88fd'] }
    RateLimit-Limit:
      description: Requests permitted per minute on this key.
      schema: { type: integer }
    RateLimit-Remaining:
      description: Whole requests left in the bucket after this one.
      schema: { type: integer }
    RateLimit-Reset:
      description: Seconds until at least one request is available. `0` means now.
      schema: { type: integer }
    Retry-After:
      description: Seconds to wait before retrying. Both SDKs honour it.
      schema: { type: integer }
    Idempotent-Replay:
      description: |
        `true` when this response replays an earlier request under the same
        `Idempotency-Key` rather than executing a new one.
      schema: { type: string, enum: ['true'] }
...
  responses:
    BadRequest:
      description: |
        The request is malformed. `error.param` names the offending field
        whenever there is one to name.
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: |
        No key, or one that is not usable. One message covers malformed,
        unknown, wrong-secret, revoked and expired: a caller holding a broken
        credential has no business learning which of the five it is.
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: |
        A valid key without the scope this route requires. Scopes are fixed when
        a key is minted — mint a new one.

        On `POST /v1/chat` and `POST /v1/reports` it is also
        `agent_not_allowed`: the key is limited to named agents, and the turn
        would run as another. Name one of the key's agents in `agent_id`.
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: |
        No such resource **for this company**. A cross-tenant id and a
        nonexistent id answer identically; anything else is an existence oracle
        over another company's data.
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: |
        Two shapes, distinguished by `error.code`. `request_in_flight` carries
        `in_flight` with the ids of the work already running;
        `idempotency_key_reuse` means the same key arrived with a different
        body.
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
        Idempotent-Replay: { $ref: '#/components/headers/Idempotent-Replay' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ConflictError' }
    BudgetExhausted:
      description: |
        The tenant is out of credit. Typed rather than a 500 on purpose: a
        programmatic caller retries a 500, and retrying a turn nobody can pay
        for is a loop that stops only when someone notices.
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    TooLarge:
      description: The request body is over this deployment's cap.
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RateLimited:
      description: Too many requests on this key.
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
        Retry-After: { $ref: '#/components/headers/Retry-After' }
        RateLimit-Limit: { $ref: '#/components/headers/RateLimit-Limit' }
        RateLimit-Remaining: { $ref: '#/components/headers/RateLimit-Remaining' }
        RateLimit-Reset: { $ref: '#/components/headers/RateLimit-Reset' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ServerError:
      description: |
        Our fault. Quote `error.request_id` — it is the string that finds this
        attempt in our logs.
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unavailable:
      description: |
        Either the whole public API is switched off (`api_disabled`, with
        `Retry-After`) or one subsystem this route needs is not configured on
        this deployment. The routes exist and are coming back, which is why this
        is not a 404.
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
        Retry-After: { $ref: '#/components/headers/Retry-After' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    ErrorDetail:
      type: object
      description: |
        Switch on `type`. `code` is the specific reason within that class, and
        only `message` is meant for a human to read.
      required: [type, code, message]
      properties:
        type:
          type: string
          description: The coarse class. The vocabulary is closed.
          enum:
            - invalid_request
            - authentication
            - permission
            - not_found
            - rate_limit
            - budget_exhausted
            - server
        code:
          type: string
          description: |
            The specific reason — `insufficient_scope`, `spec_too_large`,
            `thread_not_found`, and so on. New codes are added over time; treat
            an unknown one as its `type`.
          examples: ['insufficient_scope']
        message:
          type: string
          description: A sentence for a person. Do not switch on it.
        param:
          type: string
          description: |
            The field or header that was wrong, when there is one. Its presence
            means something — an integrator reading `"param": "rows"` fixes
            their request without reading prose.
        request_id:
          type: string
          description: The same value as the `X-Request-Id` header.
      additionalProperties: false
    Error:
      type: object
      description: |
        Every failure under `/v1` has this shape. A bare `{"error": "…"}` string
        anywhere here is a defect.
      required: [error]
      properties:
        error: { $ref: '#/components/schemas/ErrorDetail' }
      additionalProperties: false
    ConflictError:
      type: object
      description: |
        A 409. `in_flight` is present only for `request_in_flight`, and its
        contents are whatever ids identify the work — `{thread_id, run_id,
        started_at}` for a chat turn, `{report_id, status}` for a report.
      required: [error]
      properties:
        error: { $ref: '#/components/schemas/ErrorDetail' }
        in_flight:
          type: object
          description: The ids of the request that is still running.
          additionalProperties: true
      additionalProperties: false
...
    DocumentFormat:
      type: string
      description: |
        `mp4` is the same report spec as a silent 1080p video. It is rendered by
        a separate service and takes minutes rather than milliseconds, so every
        door that produces one is asynchronous: `POST /v1/reports/render` with
        `format: mp4` answers `202` whatever `Accept` says, and the file is
        collected the same three ways an agentic report is. It requires a
        report that makes an argument — a `kpi_row` or a `chart` — because a
        record such as an invoice cannot be scanned in a video. A deployment
        with no render service refuses it with `format_unavailable` and serves
        every other format as before.

        `carousel` is the same report spec as one to ten JPEG slides for a
        social feed, drawn by the same render service, and it follows every
        rule above: asynchronous, refused where there is no render service.
        The document is a zip of the pages, a `caption.txt` and a
        `carousel.json` manifest (caption, hashtags, one alt text per slide);
        the spec may carry `social.caption` (≤ 2 200 characters),
        `social.hashtags` (≤ 30) and `social.size`. A spec whose sections make
        more than ten slides is refused with the count, not truncated.

        **One slide is a whole post.** A promotion or a launch is one image:
        one `promo` section for a product with a price, or one `hero` section
        for anything else, and nothing else. Such a post is an announcement rather than an
        argument, so it is the one carousel exempt from the "argument about
        data rather than a record" rule above: a `hero` satisfies it where a
        `kpi_row` or a `chart` otherwise would. An invoice still cannot be a
        carousel, because a record carries neither.
      enum: [pdf, pptx, xlsx, csv, mp4, carousel]
    Document:
      type: object
      description: One generated file.
      required: [id, object, filename, format, size_bytes, source, created_at]
      properties:
        id: { type: string, format: uuid }
        object: { type: string, const: document }
        filename: { type: string }
        format: { $ref: '#/components/schemas/DocumentFormat' }
        size_bytes: { type: integer }
        source:
          type: string
          description: |
            Which door produced it. `api` for either `/v1` report route, `agent`
            for the `generate_document` tool inside a dashboard turn.
          enum: [agent, api]
        thread_id:
          type: string
          format: uuid
          description: |
            Absent for a document the render door produced — it has no
            conversation behind it.
        created_at: { type: string, format: date-time }
        download_url:
          type: string
          format: uri
          description: |
            Presigned and short-lived, re-issued on every read rather than
            stored. Absent on list responses and when the object store could not
            be reached.
        expires_at: { type: string, format: date-time }
      additionalProperties: false
...
    CreateReportRequest:
      type: object
      description: |
        Send `user_ref` **or** `thread_id`: the first names who the report is
        for, the second continues a conversation that already knows.
      required: [prompt]
      properties:
        prompt:
          type: string
          description: What the report should contain, in words.
          examples: ['Monthly revenue for 2024 with a bar chart by month and a short commentary.']
        format:
          allOf: [{ $ref: '#/components/schemas/DocumentFormat' }]
          description: |
            Defaults to `pdf`. **`mp4` is refused here** with
            `format_not_supported_here`: a video is rendered after the turn that
            asked for it has ended, so this report would complete before the
            file existed. Post the spec to `POST /v1/reports/render` instead.
        user_ref:
          type: string
          description: |
            Your own identifier for the person this is on behalf of. It is what
            makes the spend attributable in the tenant's usage reporting.
        thread_id:
          type: string
          format: uuid
          description: Continue an existing `api` conversation instead of starting one.
        agent_id:
          type: string
          format: uuid
          description: |
            Which of the workspace's agents writes the report. Omit it for the
            default; list them with `GET /v1/agents`. Same field, same rules and
            same errors as `POST /v1/chat` — a report is an agent turn that ends
            in a document.
        callback_url:
          type: string
          format: uri
          description: |
            Receives the signed `report.completed` body. Optional — most callers
            poll or stream. A private-network target is refused unless the
            deployment allows one.
        locale:
          type: string
          description: '`id` or `en`. Decides separators, month names and magnitude words.'
        currency:
          type: string
          description: An ISO 4217 code.
      additionalProperties: false
    Report:
      type: object
      description: |
        One shape for both doors and for the poll route, so a caller writes the
        collection path once whether the job came from a prompt or from a spec
        that ran long.
      required: [id, object, status, kind, format, created_at]
      properties:
        id: { type: string, format: uuid }
        object: { type: string, const: report }
        status:
          type: string
          enum: [queued, running, completed, failed]
        kind:
          type: string
          description: '`agentic` for a prompt; `render` for a spec that outran the synchronous window.'
          enum: [agentic, render]
        format: { $ref: '#/components/schemas/DocumentFormat' }
        thread_id:
          type: string
          format: uuid
          description: |
            The conversation behind an agentic report — continue it through
            `POST /v1/chat`. Absent for a render job, which has none.
        document: { $ref: '#/components/schemas/Document' }
        error:
          type: string
          description: Why a `failed` report failed.
        created_at: { type: string, format: date-time }
        request_id: { type: string }
      additionalProperties: false
...
    Thread:
      type: object
      description: |
        A conversation. Deliberately not the internal thread record, which
        carries a phone number, a Discord user id and two Lark keys — none of
        which an `api` thread has, and all of which would become part of a
        published contract the moment one was serialized.
      required: [id, object, last_message_at, created_at]
      properties:
        id: { type: string, format: uuid }
        object: { type: string, const: thread }
        user_ref: { type: string }
        title: { type: string }
        summary:
          type: string
          description: The rolling topic summary the thread resolver keeps.
        last_message_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
      additionalProperties: false
    ThreadPage:
      type: object
      required: [data, has_more]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Thread' }
        has_more: { type: boolean }
        next_cursor: { type: string }
      additionalProperties: false
...